Refactoring, move two methods to ApdbCassandraAdmin class.
