+++ python/lsst/dax/apdb/cassandra/apdbCassandraAdmin.py
@@ -23,8 +23,10 @@ from __future__ import annotations
 
 __all__ = ["ApdbCassandraAdmin"]
 
+import dataclasses
 import itertools
 import logging
+import warnings
 from collections import defaultdict
 from collections.abc import Iterable, Mapping
 from typing import TYPE_CHECKING
@@ -32,10 +34,18 @@ from typing import TYPE_CHECKING
 from lsst.sphgeom import LonLat, UnitVector3d
 from lsst.utils.iteration import chunk_iterable
 
+try:
+    import cassandra
+except ImportError:
+    pass
+
 from ..apdbAdmin import ApdbAdmin, DiaForcedSourceLocator, DiaObjectLocator, DiaSourceLocator
+from ..apdbSchema import ApdbTables
 from ..monitor import MonAgent
 from ..timer import Timer
-from .cassandra_utils import execute_concurrent
+from .cassandra_utils import execute_concurrent, quote_id
+from .config import ApdbCassandraConfig
+from .sessionFactory import SessionContext
 
 if TYPE_CHECKING:
     from .apdbCassandra import ApdbCassandra
@@ -43,6 +53,22 @@ if TYPE_CHECKING:
 _LOG = logging.getLogger(__name__)
 
 _MON = MonAgent(__name__)
+
+
+@dataclasses.dataclass
+class DatabaseInfo:
+    """Collection of information about a specific database."""
+
+    name: str
+    """Keyspace name."""
+
+    permissions: dict[str, set[str]] | None = None
+    """Roles that can access the database and their permissions.
+
+    `None` means that authentication information is not accessible due to
+    system table permissions. If anonymous access is enabled then dictionary
+    will be empty but not `None`.
+    """
 
 
 class ApdbCassandraAdmin(ApdbAdmin):
@@ -60,6 +86,84 @@ class ApdbCassandraAdmin(ApdbAdmin):
     def _timer(self, name: str, *, tags: Mapping[str, str | int] | None = None) -> Timer:
         """Create `Timer` instance given its name."""
         return Timer(name, _MON, _LOG, tags=tags)
+
+    @classmethod
+    def list_databases(cls, host: str) -> Iterable[DatabaseInfo]:
+        """Return the list of keyspaces with APDB databases.
+
+        Parameters
+        ----------
+        host : `str`
+            Name of one of the hosts in Cassandra cluster.
+
+        Returns
+        -------
+        databases : `~collections.abc.Iterable` [`DatabaseInfo`]
+            Information about databases that contain APDB instance.
+        """
+        # For DbAuth we need to use database name "*" to try to match any
+        # database.
+        config = ApdbCassandraConfig(contact_points=(host,), keyspace="*")
+        with SessionContext(config) as session:
+            # Get names of all keyspaces containing DiaSource table
+            table_name = ApdbTables.DiaSource.table_name()
+            query = "select keyspace_name from system_schema.tables where table_name = %s ALLOW FILTERING"
+            result = session.execute(query, (table_name,))
+            keyspaces = [row[0] for row in result.all()]
+
+            if not keyspaces:
+                return []
+
+            # Retrieve roles for each keyspace.
+            template = ", ".join(["%s"] * len(keyspaces))
+            query = (
+                "SELECT resource, role, permissions FROM system_auth.role_permissions "
+                f"WHERE resource IN ({template}) ALLOW FILTERING"
+            )
+            resources = [f"data/{keyspace}" for keyspace in keyspaces]
+            try:
+                result = session.execute(query, resources)
+                # If anonymous access is enabled then result will be empty,
+                # set infos to have empty permissions dict in that case.
+                infos = {keyspace: DatabaseInfo(name=keyspace, permissions={}) for keyspace in keyspaces}
+                for row in result:
+                    _, _, keyspace = row[0].partition("/")
+                    role: str = row[1]
+                    role_permissions: set[str] = set(row[2])
+                    infos[keyspace].permissions[role] = role_permissions  # type: ignore[index]
+            except cassandra.Unauthorized as exc:
+                # Likely that access to role_permissions is not granted for
+                # current user.
+                warnings.warn(
+                    f"Authentication information is not accessible to current user - {exc}", stacklevel=2
+                )
+                infos = {keyspace: DatabaseInfo(name=keyspace) for keyspace in keyspaces}
+
+            # Would be nice to get size estimate, but this is not available
+            # via CQL queries.
+            return infos.values()
+
+    @classmethod
+    def delete_database(cls, host: str, keyspace: str, *, timeout: int = 3600) -> None:
+        """Delete APDB database by dropping its keyspace.
+
+        Parameters
+        ----------
+        host : `str`
+            Name of one of the hosts in Cassandra cluster.
+        keyspace : `str`
+            Name of keyspace to delete.
+        timeout : `int`, optional
+            Timeout for delete operation in seconds. Dropping a large keyspace
+            can be a long operation, but this default value of one hour should
+            be sufficient for most or all cases.
+        """
+        # For DbAuth we need to use database name "*" to try to match any
+        # database.
+        config = ApdbCassandraConfig(contact_points=(host,), keyspace="*")
+        with SessionContext(config) as session:
+            query = f"DROP KEYSPACE {quote_id(keyspace)}"
+            session.execute(query, timeout=timeout)
 
     def apdb_part(self, ra: float, dec: float) -> int:
         # docstring is inherited from a base class
